+++ src/test/java/site/nomoreparties/stellarburgers/CreateOrderTest.java
@@ -6,6 +6,9 @@ import io.restassured.response.ValidatableResponse;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import site.nomoreparties.stellarburgers.steps.GeneralSteps;
+import site.nomoreparties.stellarburgers.steps.OrderSteps;
+import site.nomoreparties.stellarburgers.steps.UserSteps;
 
 import java.util.List;
 
@@ -13,25 +16,27 @@ import static org.apache.http.HttpStatus.*;
 import static org.hamcrest.CoreMatchers.*;
 
 @DisplayName("Создание заказа POST: /api/orders/")
-public class CreateOrderTest extends Steps {
-
+public class CreateOrderTest {
     private static final String BAD_REQUEST_EXPECTED_MESSAGE = "Ingredient ids must be provided";
+    private GeneralSteps generalSteps = new GeneralSteps();
+    private UserSteps userSteps = new UserSteps();
+    private OrderSteps orderSteps = new OrderSteps();
 
     @Before
     public void setUp() {
-        turnOnLogs();
-        createOrderClient();
+        generalSteps.turnOnLogs();
+        orderSteps.createOrderClient();
     }
     @Test
     @DisplayName("Тест на создание заказа с ингредиентами и с авторизацией")
     @Description("Проверка на создание заказа с авторизацией и ингредиентами")
     public void createNewOrderAuth() {
-        createUserClient();
-        createUser();
-        ValidatableResponse response = login();
-        accessToken = getAccessToken(response);
-        List<String> twoIngredients = getTwoIngredients();
-        createNewOrderAuth(twoIngredients, accessToken).assertThat().body("success", equalTo(true))
+        userSteps.createUserClient();
+        userSteps.createUser();
+        ValidatableResponse response = userSteps.login();
+        userSteps.accessToken = userSteps.getAccessToken(response);
+        List<String> twoIngredients =  orderSteps.getTwoIngredients();
+        orderSteps.createNewOrderAuth(twoIngredients, userSteps.accessToken).assertThat().body("success", equalTo(true))
                 .body("name", notNullValue())
                 .body("order",notNullValue())
                 .and()
@@ -41,8 +46,8 @@ public class CreateOrderTest extends Steps {
     @DisplayName("Тест на создание заказа с ингредиентами и без авторизации")
     @Description("Проверка на создание заказа без авторизации и c ингредиентами")
     public void createNewOrderNotAuth() {
-        List<String> twoIngredients = getTwoIngredients();
-        createNewOrderNotAuth(twoIngredients).assertThat().body("success", equalTo(true))
+        List<String> twoIngredients =  orderSteps.getTwoIngredients();
+        orderSteps.createNewOrderNotAuth(twoIngredients).assertThat().body("success", equalTo(true))
                 .body("name", notNullValue())
                 .body("order",notNullValue())
                 .and()
@@ -52,8 +57,8 @@ public class CreateOrderTest extends Steps {
     @DisplayName("Тест на создание заказа без ингредиентов")
     @Description("Проверка на создание заказа без ингредиентов и получение ошибки")
     public void createNewOrderNullIngredients() {
-        List<String> nullIngredients = getNullIngredients();
-        createNewOrderNotAuth(nullIngredients).assertThat().body("success", equalTo(false))
+        List<String> nullIngredients =  orderSteps.getNullIngredients();
+        orderSteps.createNewOrderNotAuth(nullIngredients).assertThat().body("success", equalTo(false))
                 .body("message",  equalTo(BAD_REQUEST_EXPECTED_MESSAGE))
                 .and()
                 .statusCode(SC_BAD_REQUEST);
@@ -62,12 +67,12 @@ public class CreateOrderTest extends Steps {
     @DisplayName("Тест на создание заказа с неверным хешем")
     @Description("Проверка на создание заказа неверным хешем и получение ошибки")
     public void createNewOrderInvalidIngredients() {
-        List<String> invalidIngredient = getInvalidIngredient();
-        createNewOrderNotAuth(invalidIngredient).statusCode(SC_INTERNAL_SERVER_ERROR);
+        List<String> invalidIngredient =  orderSteps.getInvalidIngredient();
+        orderSteps.createNewOrderNotAuth(invalidIngredient).statusCode(SC_INTERNAL_SERVER_ERROR);
     }
     @After
     public void deleteTestDataAndTurnOffLogs() {
-        deleteUser();
-        turnOffLogs();
+        userSteps.deleteUser();
+        generalSteps.turnOffLogs();
     }
 }
