+++ src/test/java/site/nomoreparties/stellarburgers/EditUserTest.java
@@ -6,31 +6,33 @@ import io.restassured.response.ValidatableResponse;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-
+import site.nomoreparties.stellarburgers.steps.GeneralSteps;
+import site.nomoreparties.stellarburgers.steps.UserSteps;
 
 import static org.apache.http.HttpStatus.*;
 import static org.hamcrest.CoreMatchers.equalTo;
 
 @DisplayName("Изменение данных пользователя PATCH: /api/auth/user/")
-public class EditUserTest extends Steps {
+public class EditUserTest {
 
     private static final String UNAUTHORIZED_EXPECTED_MESSAGE = "You should be authorised";
-
+    private GeneralSteps generalSteps = new GeneralSteps();
+    private UserSteps userSteps = new UserSteps();
     @Before
     public void setUp() {
-        turnOnLogs();
-        createUserClient();
+        generalSteps.turnOnLogs();
+        userSteps.createUserClient();
     }
     @Test
     @DisplayName("Изменение данных пользователя с авторизацией")
     @Description("Проверка что у авторизованного пользователя можно изменять поля")
     public void editUserAuth() {
-        createUser();
-        ValidatableResponse response = login();
-        accessToken = getAccessToken(response);
-        editAuth().assertThat().body("success", equalTo(true))
-                .body("user.email", equalTo(fakerEmail))
-                .body("user.name", equalTo(fakerName))
+        userSteps.createUser();
+        ValidatableResponse response = userSteps.login();
+        userSteps.accessToken = userSteps.getAccessToken(response);
+        userSteps.editAuth().assertThat().body("success", equalTo(true))
+                .body("user.email", equalTo(userSteps.fakerEmail))
+                .body("user.name", equalTo(userSteps.fakerName))
                 .and()
                 .statusCode(SC_OK);
     }
@@ -38,14 +40,14 @@ public class EditUserTest extends Steps {
     @DisplayName("Изменение данных пользователя без авторизацией")
     @Description("Проверка что у неавторизованного пользователя нельзя изменять поля")
     public void editUserNotAuth() {
-        editNotAuth().assertThat().body("success", equalTo(false))
+        userSteps.editNotAuth().assertThat().body("success", equalTo(false))
                 .body("message", equalTo(UNAUTHORIZED_EXPECTED_MESSAGE))
                 .and()
                 .statusCode(SC_UNAUTHORIZED);
     }
     @After
     public void deleteTestDataAndTurnOffLogs() {
-        deleteUser();
-        turnOffLogs();
+        userSteps.deleteUser();
+        generalSteps.turnOffLogs();
     }
 }
